requirements.txt
streamlit==1.28.1
pdf2image==1.16.3
Pillow==10.0.1
ultralytics==8.0.196
huggingface-hub==0.19.4
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78